first pass of cleaning repo
