+++ .gitignore
@@ -5,6 +5,9 @@
 
 libraries/
 
+connect.sh
+db_config.py
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
