+++ config_files/paths.py
@@ -1,0 +1,5 @@
+AVA_IMAGE_PATH = '/media/matt/New Volume/ava/ava-compressed/images/'
+MISSOURIAN_IMAGE_PATH = '/media/matt/New Volume/missourian_pics/2017/Fall/Dump/Compton, Nate/20170824_MUfootballpractice_NC'
+AVA_QUALITY_LABELS_FILE = '/media/matt/New Volume/ava/AVA.txt'
+AVA_CONTENT_LABELS_FILE = '/media/matt/New Volume/ava/tags.txt'
+MODEL_STORAGE_PATH = '/College/research/RJI-Photo/models/'
