+++ config_files/xmp_to_ratings.py
@@ -1,0 +1,10 @@
+XMP_AVA_MAP = {
+    1: 10,
+    2: 9,
+    3: 7,
+    4: 6,
+    5: 5,
+    6: 4,
+    7: 2,
+    8: 1
+}
