+++ database/sql/insert.sql
@@ -1,0 +1,5 @@
+INSERT INTO rji.models (model_name, epochs, start_lr, m_type, num_ouputs, loss_fn) VALUES ? ? ? ? ? ?;
+
+INSERT INTO rji.photos (photo_fname, ranking) VALUES ? ?;
+
+INSERT INTO rji.losses (model_id, epoch, train_loss, validation_loss) VALUES ? ? ? ?;
