+++ database/sql/select.sql
@@ -1,0 +1,2 @@
+SELECT model_id FROM rji.models ORDER BY rji.model_id DESC LIMIT 1;
+
