--- scripts/sql/select_paths.sql
@@ -1,4 +1,0 @@
-SELECT photo_path, cluster_number
-        FROM cluster_results
-        WHERE cluster_session_id = 19
-
